Add docker-compose; add flask; add-mongodb
