--- requirements.txt
@@ -1,6 +1,0 @@
-certifi==2022.6.15
-charset-normalizer==2.1.0
-idna==3.3
-pyTelegramBotAPI==4.6.1
-requests==2.28.1
-urllib3==1.26.11
