+++ src/.gitignore
@@ -1,2 +1,3 @@
 .bot_venv/
 credentials.py
+.env.example
