+++ src/Dockerfile
@@ -1,7 +1,7 @@
 FROM python:3
-ENV PYTHONUNBUFFERED 1
 WORKDIR /app
 COPY requirements.txt .
 RUN pip install -r requirements.txt
 COPY . .
+ENV PYTHONPATH ./
 CMD [ "python", "./main.py" ]
